Update: Fixed multiple modifiers needing to be applied
Dayside Farm planetary modifier runs into issues if the owner has both the Mastery of Nature ascension perk, and the tidal lock knowledge.

I've had to hardcode the logic due to how Paradox calculate mult values. I'm basing the maths on the idea that the tidal lock knowledge changes the "base" amount of generator districts from 6 to 12. The Mastery of Nature ascension perk would then apply the +50% increase in resource districts to the new "base" amount, resulting in 18 generator districts.

The logic is now the following:
- Dayside Farm (No = Mastery of Nature, No = Tidal Lock Knowledge): +6 generator districts & +30% energy production modifier
- Dayside Farm (Yes = Mastery of Nature, No = Tidal Lock Knowledge): +9 generator districts & +30% energy production modifier
- Dayside Farm (No = Mastery of Nature, Yes = Tidal Lock Knowledge): +12 generator districts & +60% energy production modifier
- Dayside Farm (Yes = Mastery of Nature, Yes = Tidal Lock Knowledge): +18 generator districts & +60% energy production modifier